Restructure sound package
