+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/sound/file/SoundFile.kt
@@ -1,4 +1,4 @@
-package liltojustice.trueadaptivemusic.client.sound
+package liltojustice.trueadaptivemusic.client.sound.file
 
 import java.io.InputStream
