+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/sound/playable/PlayableSoundEvent.kt
@@ -1,4 +1,4 @@
-package liltojustice.trueadaptivemusic.client.sound
+package liltojustice.trueadaptivemusic.client.sound.playable
 
 import net.minecraft.client.sound.PositionedSoundInstance
 import net.minecraft.client.sound.SoundInstance
